backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
google-cloud-bigquery==3.13.0
google-auth==2.25.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
httpx==0.25.2
google-genai>=0.3.0
termcolor==2.3.0
tabulate==0.9.0
slack_sdk==3.29.0
APScheduler==3.10.4
